# perf: reference management optimizations